requirements.txt
ailment==9.2.123
angr==9.2.123
anyio==4.6.2.post1
archinfo==9.2.123
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bitarray==2.9.3
bitstring==4.2.3
bleach==6.1.0
cachetools==5.5.0
capstone==5.0.3
cart==1.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
claripy==9.2.123
cle==9.2.123
colored-traceback==0.4.2
comm==0.2.2
contourpy==1.3.0
CppHeaderParser==2.7.4
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
fastjsonschema==2.20.0
fonttools==4.54.1
fqdn==1.5.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.28.0
isoduration==20.11.0
itanium-demangler==1.1
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
lief==0.15.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
mulpyplexer==0.9
nampa==0.1.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.2
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
pefile==2024.8.26
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plumbum==1.9.0
ply==3.11
prometheus_client==0.21.0
prompt_toolkit==3.0.48
protobuf==5.28.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pwntools==4.13.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydot==3.0.2
pyelftools==0.31
pyformlang==1.0.10
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.2.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyvex==9.2.123
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.2
ROPGadget==7.5
rpds-py==0.20.0
rpyc==6.0.1
Send2Trash==1.8.3
setuptools==75.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.3
terminado==0.18.1
tinycss2==1.3.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
unicorn==2.0.1.post1
unique_log_filter==0.1.0
unix-ar==0.2.1
uri-template==1.3.0
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
z3-solver==4.13.0.0
zstandard==0.23.0
